feat: add CustomerMapper and related DTOs
- Introduce `CustomerMapper` for converting `CustomerRequestDTO` to `CustomerModel`
- Create `CustomerRequestDTO`, `CustomerResponseDTO`, and `CustomerUpdateDTO` for structured data transfer
- Adjust enums and repository IDs to use `Long` type
- Correct package paths for consistency within the project
- Enhance `CustomerService` with CRUD operations leveraging new DTOs and `CustomerMapper`

+++ src/main/java/com/main/numOps/domain/did/dtos/ActivateDidRequest.java
@@ -9,7 +9,7 @@ import java.util.List;
 public record ActivateDidRequest(
 
         @NotEmpty(message = "A lista de números é obrigatória")
-        List<@NotNull(message = "Número inválido") Integer> numeros,
+        List<@NotNull(message = "Número inválido") Long> numeros,
 
         @NotBlank(message = "Cliente não pode ser vazio")
         String razao,
